First iteration
Although this iteration does not contain a complete walking skeleton (it
does not have a console app to run the project) the architecture is
based in the end to end tests:

* Complete failure when Player does not place a bet
* And Player loses even when she places a bet

Most classes does not make nothing special right now and the game logic
is not implemented at all (there are no cards nor hands nor dealer) and
this will be implemented as soon another tests appear.

The thing that bothers me most is that this initial organization may
complicate the insertion of the Dealer and its association in game
(since it should shuffle cards and deal cards at some point) and I hope
that it does not get in the way in the future.